+++ MiamIOSFramework/MiamIOSFramework/ui/BasketPreview/BasketPreviewHeader.swift
@@ -34,20 +34,17 @@ struct BasketPreviewHeader: View {
     }
 
 
-    let decreaseGuestsCount: () -> Void
-    let increaseGuestsCount: () -> Void
+    let updateGuest: (Int) -> Void
     let goToDetail: () -> Void
 
     var body: some View {
         if (Template.sharedInstance.basketPreviewHeaderTemplate != nil) {
-            Template.sharedInstance.basketPreviewHeaderTemplate!(basketTitle, pictureURL, basketDescription, price, numberOfGuests, price, decreaseGuestsCount, increaseGuestsCount)
+            Template.sharedInstance.basketPreviewHeaderTemplate!(basketTitle, pictureURL, basketDescription, price, numberOfGuests, price, updateGuest)
         } else {
             VStack(alignment: .leading) {
                 HStack {
                     if hasPicture {
-                        AsyncImage(url: pictureURL!, placeholder: {
-                                ProgressView()
-                        }, height: 120.0).frame(width:120, height: 120, alignment: .topLeading).cornerRadius(12.0)
+                        AsyncImage(url: pictureURL!, height: 120.0).frame(width:120, height: 120, alignment: .topLeading).cornerRadius(12.0)
                     }
 
                     VStack (alignment: .leading) {
@@ -75,10 +72,8 @@ struct BasketPreviewHeader: View {
                     Text(formattedPrice).foregroundColor(Color.miamColor(.primary)).fontWeight(.bold)
                     Spacer()
 
-                    CounterView(count: numberOfGuests) {
-                        increaseGuestsCount()
-                    } decrease: {
-                        decreaseGuestsCount()
+                    CounterView(count: numberOfGuests)  { guestCount in
+                        updateGuest(guestCount)
                     }
                 }
             }.padding([.leading, .trailing], 16.0)