+++ androidSDK/src/main/java/com/miam/kmm_miam_sdk/android/ui/components/basketPreview/EntryLine.kt
@@ -8,10 +8,6 @@ import androidx.compose.material.IconButton
 import androidx.compose.material.Surface
 import androidx.compose.material.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
@@ -31,13 +27,13 @@ import com.miam.kmm_miam_sdk.android.theme.Typography.body
 import com.miam.kmm_miam_sdk.android.theme.Typography.bodyBold
 import com.miam.kmm_miam_sdk.android.theme.Typography.bodySmallBold
 import com.miam.kmm_miam_sdk.android.ui.components.basketPreview.customization.BasketPreviewColor.BPPLDescriptionColor
-import com.miam.kmm_miam_sdk.android.ui.components.basketPreview.customization.BasketPreviewImage.delete
+import com.miam.kmm_miam_sdk.android.ui.components.basketPreview.customization.BasketPreviewImage
 import com.miam.kmm_miam_sdk.android.ui.components.basketPreview.customization.BasketPreviewImage.swap
 import com.miam.kmm_miam_sdk.android.ui.components.basketPreview.customization.BasketPreviewText
 import com.miam.kmm_miam_sdk.android.ui.components.basketPreview.subcomponent.MultipleRecipeTag
 import com.miam.kmm_miam_sdk.android.ui.components.common.Clickable
 import com.miam.kmm_miam_sdk.android.ui.components.counter.Counter
-import com.miam.kmm_miam_sdk.android.ui.components.price.Price
+import com.miam.kmm_miam_sdk.android.ui.components.price.SimplePrice
 import java.util.*
 
 @ExperimentalCoilApi
@@ -47,19 +43,9 @@ fun EntryLine(
     vmBasketPreview: BasketPreviewViewModel,
     goToItemSelector: () -> Unit
 ) {
-    val productName =
-        entry.title.substring(0, 1).uppercase(Locale.getDefault()) + entry.title.substring(1)
-            .lowercase(Locale.getDefault())
+    val productName = entry.title.substring(0, 1).uppercase(Locale.getDefault()) + entry.title.substring(1).lowercase(Locale.getDefault())
     val description = entry.bplDescription[0]
     val sharingCount = entry.inlineTag
-    var count by remember { mutableStateOf(entry.count) }
-    val rememberedEntry by remember { mutableStateOf(entry) }
-
-    if (rememberedEntry.id != entry.id) {
-        // object is reused reset remembered count
-        count = entry.count
-    }
-
 
     fun delete() {
         vmBasketPreview.removeBasketEntry(entry.record as BasketEntry)
@@ -70,18 +56,12 @@ fun EntryLine(
         goToItemSelector()
     }
 
-    fun increaseQty() {
-        count++
-        vmBasketPreview.updateBasketEntry(entry.record as BasketEntry, count)
-    }
-
-    fun decreaseQty() {
-        if (count == 1) {
+    fun onQuantityChanged(newQuantity: Int) {
+        if (newQuantity == 0) {
             vmBasketPreview.removeBasketEntry(entry.record as BasketEntry)
-        } else {
-            count--
-            vmBasketPreview.updateBasketEntry(entry.record as BasketEntry, count)
+            return
         }
+        vmBasketPreview.updateBasketEntry(entry.record as BasketEntry, newQuantity)
     }
 
     if (Template.basketPreviewProductLine != null) {
@@ -90,73 +70,98 @@ fun EntryLine(
                 productName,
                 description,
                 entry.picture,
-                count,
+                entry.count,
                 sharingCount.toString(),
                 entry.price,
-                (entry.record as BasketEntry).relationships!!.items!!.data.size,
+                (entry.record as BasketEntry).itemsCountOrZero,
                 { delete() },
                 { replace() },
-                { increaseQty() },
-                { decreaseQty() })
+                { newQuantity -> onQuantityChanged(newQuantity) },
+                null
+            )
         }
     } else {
+        BasketPreviewProductLine(
+            productName,
+            description,
+            entry.picture,
+            entry.count,
+            sharingCount.toString(),
+            entry.price,
+            (entry.record as BasketEntry).itemsCountOrZero,
+            { delete() },
+            { replace() }
+        ) { newQuantity -> onQuantityChanged(newQuantity) }
+    }
+}
 
-        Spacer(modifier = Modifier.padding(vertical = 8.dp))
-        Column {
-            Row(
-                modifier = Modifier.fillMaxWidth(),
-                verticalAlignment = Alignment.Top,
-                horizontalArrangement = Arrangement.SpaceBetween
+@Composable
+fun BasketPreviewProductLine(
+    productName: String,
+    description: String,
+    picture: String,
+    quantity: Int,
+    sharingCount: String,
+    price: String,
+    itemsCount: Int,
+    delete: () -> Unit,
+    replace: () -> Unit,
+    onQuantityChanged: (Int) -> Unit
+) {
+    Spacer(modifier = Modifier.padding(vertical = 8.dp))
+    Column {
+        Row(
+            modifier = Modifier.fillMaxWidth(),
+            verticalAlignment = Alignment.Top,
+            horizontalArrangement = Arrangement.SpaceBetween
+        ) {
+            Image(
+                painter = rememberImagePainter(picture),
+                contentDescription = null,
+                contentScale = ContentScale.Crop,
+                modifier = Modifier
+                    .height(80.dp)
+                    .width(80.dp)
+                    .clip(RoundedCornerShape(8.dp))
+            )
+            Spacer(modifier = Modifier.padding(horizontal = 8.dp))
+            Column {
+                Text(
+                    text = productName,
+                    style = bodyBold
+                )
+                Spacer(modifier = Modifier.padding(vertical = 2.dp))
+                Text(
+                    text = description,
+                    color = BPPLDescriptionColor,
+                    style = bodySmallBold,
+                    modifier = Modifier.widthIn(200.dp, 200.dp)
+                )
+                Spacer(modifier = Modifier.padding(vertical = 2.dp))
+                MultipleRecipeTag(sharingCount)
+            }
+            IconButton(
+                modifier = Modifier.size(30.dp),
+                onClick = { delete() }
             ) {
                 Image(
-                    painter = rememberImagePainter(entry.picture),
-                    contentDescription = null,
-                    contentScale = ContentScale.Crop,
-                    modifier = Modifier
-                        .height(80.dp)
-                        .width(80.dp)
-                        .clip(RoundedCornerShape(8.dp))
+                    painter = painterResource(BasketPreviewImage.delete),
+                    contentDescription = "delete"
                 )
-                Spacer(modifier = Modifier.padding(horizontal = 8.dp))
-                Column {
-                    Text(
-                        text = productName,
-                        style = bodyBold
-                    )
-                    Spacer(modifier = Modifier.padding(vertical = 2.dp))
-                    Text(
-                        text = description,
-                        color = BPPLDescriptionColor,
-                        style = bodySmallBold,
-                        modifier = Modifier.widthIn(200.dp, 200.dp)
-                    )
-                    Spacer(modifier = Modifier.padding(vertical = 2.dp))
-                    MultipleRecipeTag(sharingCount)
-                }
-                IconButton(
-                    modifier = Modifier.size(30.dp),
-                    onClick = { delete() }
-                ) {
-                    Image(
-                        painter = painterResource(delete),
-                        contentDescription = "delete"
-                    )
-                }
             }
+        }
 
-            Column(Modifier.fillMaxWidth()) {
-                Spacer(modifier = Modifier.padding(vertical = 4.dp))
-                EntryPriceAndActionRow(
-                    (entry.record as BasketEntry).relationships!!.items!!.data.size,
-                    entry.price.toDouble(),
-                    count,
-                    ::increaseQty,
-                    ::decreaseQty,
-                    ::replace
-                )
-                Spacer(modifier = Modifier.padding(vertical = 4.dp))
-                Divider(Modifier.fillMaxWidth(), color = Color.LightGray)
-            }
+        Column(Modifier.fillMaxWidth()) {
+            Spacer(modifier = Modifier.padding(vertical = 4.dp))
+            EntryPriceAndActionRow(
+                itemsCount,
+                price.toDouble(),
+                quantity,
+                { onQuantityChanged(it) },
+                { replace() },
+            )
+            Spacer(modifier = Modifier.padding(vertical = 4.dp))
+            Divider(Modifier.fillMaxWidth(), color = Color.LightGray)
         }
     }
 }
@@ -166,9 +171,8 @@ fun EntryLine(
 fun EntryPriceAndActionRow(
     itemsCount: Int,
     price: Double,
-    count: Int,
-    increaseQty: () -> Unit,
-    decreaseQty: () -> Unit,
+    currentEntryCount: Int,
+    onCounterChanged: (counterValue: Int) -> Unit,
     replace: () -> Unit
 ) {
     Row(
@@ -205,17 +209,15 @@ fun EntryPriceAndActionRow(
             Modifier.weight(1f)
         )
         Box(modifier = Modifier.padding(end = 16.dp)) {
-            Price(
-                price = price,
-                isTotalPrice = true
-            )
+            SimplePrice(price = price)
         }
         Counter(
-            count = count,
-            increase = { increaseQty() },
-            decrease = { decreaseQty() },
+            initialCount = currentEntryCount,
+            onCounterChanged = onCounterChanged,
             lightMode = true,
-            isDisable = false
+            isDisable = false,
+            minValue = 0,
+            maxValue = 99
         )
     }
 }
@@ -223,5 +225,5 @@ fun EntryPriceAndActionRow(
 @Preview
 @Composable
 fun EntryPriceAndActionRowPreview() {
-    EntryPriceAndActionRow(2, 14.90, 4, {}, {}, {})
+    EntryPriceAndActionRow(2, 14.90, 4, {}, {})
 }