+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/component/recipe/RecipeViewModel.kt
@@ -14,6 +14,8 @@ import com.miam.kmmMiamCore.miam_core.model.Recipe
 import com.miam.kmmMiamCore.miam_core.model.SuggestionsCriteria
 import com.miam.kmmMiamCore.services.Analytics
 import kotlinx.coroutines.CoroutineExceptionHandler
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.flow.MutableSharedFlow
 import kotlinx.coroutines.flow.collect
@@ -72,7 +74,7 @@ open class RecipeViewModel(val routerVM: RouterOutletViewModel): BaseViewModel<R
                 handleGLChange(it)
             }
         }
-        launch(coroutineHandler) {
+        CoroutineScope(Dispatchers.Default).launch(coroutineHandler) {
             listenguestSubjectChanges()
         }
         setState { copy(likeIsEnable = userStore.state.value.likeIsEnable) }
@@ -119,14 +121,6 @@ open class RecipeViewModel(val routerVM: RouterOutletViewModel): BaseViewModel<R
         }
     }
 
-    fun increaseGuest() {
-        updateGuest(currentState.guest + 1)
-    }
-
-    fun decreaseGuest() {
-        updateGuest(currentState.guest - 1)
-    }
-
     fun updateGuest(nbGuest: Int) {
         // reduce guest between min and max
         var boundedGuests = max(MIN_GUESTS, nbGuest)
@@ -136,8 +130,9 @@ open class RecipeViewModel(val routerVM: RouterOutletViewModel): BaseViewModel<R
 
     private fun addOrAlterRecipe(): Job {
         val action = GroceriesListAction.AlterRecipeList(recipe!!.id, uiState.value.guest)
+        setState { copy(guestUpdating = true) }
         val job = groceriesListStore.dispatch(action)
-        setState { copy(isInCart = true) }
+        job.invokeOnCompletion { setState { copy(guestUpdating = false) } }
         return job
     }