+++ miamCore/src/commonMain/kotlin/com/miam/kmmMiamCore/miam_core/data/repository/PricingRepository.kt
@@ -3,5 +3,5 @@ package com.miam.kmmMiamCore.miam_core.data.repository
 import com.miam.kmmMiamCore.miam_core.model.Pricing
 
 interface PricingRepository {
-    suspend fun getRecipePrice(recipeId: String, posId: Int): Pricing
+    suspend fun getRecipePrice(recipeId: String, posId: Int, serves: Int?): Pricing
 }