+++ changelog.md
@@ -1,5 +1,13 @@
-# 3.2.1
-- [FIX][AND] Add missing templates in like button 
+# 3.3.0
+- [FEA][COR] Clean counter usage
+- [FEA][COR] Optimize price component
+- [FIX][AND] Add missing templates in like button and preferences search
+- [FIX][IOS] Unwrapp AsyncImage URL and replace placeholder 
+- [FIX][IOS] Add function preferenceTapped to catalogViewToolbarTemplate
+- [FIX][AND] Fix catalog Empty state background color not overridable
+- [FEA][AND] add entrie line templates in basket preview 
+- [FIX][AND] Add a floating template to category list alow to customize container modifier
+
 # 3.2.0
 - [FIX][AND] Add missing templates in myMeal and item selector
 - [FIX][AND] Add missing parameter in catalog header template